requirements.txt
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
branca==0.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Fiona==1.8.22
folium==0.13.0
fonttools==4.38.0
geopandas==0.12.1
idna==3.4
ipykernel==6.16.2
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.3
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
pytz==2022.5
pywin32==304
pyzmq==24.0.1
requests==2.28.1
Shapely==1.8.5.post1
six==1.16.0
stack-data==0.6.0
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
